config.ts - move to inside of bdsx/
