+++ bdsx/launcher.ts
@@ -1,4 +1,4 @@
-import { Config } from "../config";
+import { Config } from "./config";
 import { createAbstractObject } from "./abstractobject";
 import { asmcode } from "./asm/asmcode";
 import { asm, Register } from "./assembler";
